+++ .vscode/settings.json
@@ -1,0 +1,10 @@
+{
+	"launch": {
+
+		"configurations": [],
+		"compounds": []
+	},
+	"workbench.colorCustomizations": {
+		"editorError.foreground": "#00ff00"		
+	}
+}
