+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "^7.2.1",
     "classnames": "^2.2.6",
     "json-update": "^5.3.0",
+    "node-sass": "^4.14.1",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-alert": "^7.0.2",
@@ -40,12 +41,14 @@
     ]
   },
   "devDependencies": {
+    "css-loader": "^4.2.2",
     "eslint": "^6.6.0",
     "eslint-config-airbnb": "^18.2.0",
     "eslint-plugin-import": "^2.22.0",
     "eslint-plugin-jsx-a11y": "^6.3.1",
     "eslint-plugin-react": "^7.20.6",
     "eslint-plugin-react-hooks": "^4.1.0",
-    "redux-devtools": "^3.6.1"
+    "redux-devtools": "^3.6.1",
+    "style-loader": "^1.2.1"
   }
 }
