+++ src/Components/Footer/Footer.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import FilterLink from '../../Containers/FilterLink';
 import { VisibilityFilters } from '../../Actions/index';
-import './Footer.css';
+import styles from './Footer.module.css';
 
 const Footer = () => (
-  <div className="footer">
+  <div className={styles.footer}>
     <span>Show: </span>
     <FilterLink filter={VisibilityFilters.ShowAll}>All</FilterLink>
     <FilterLink filter={VisibilityFilters.ShowActive}>Active</FilterLink>
