+++ src/Components/TodoItem/TodoItem.jsx
@@ -1,15 +1,19 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import './TodoItem.css';
+import styles from './TodoItem.module.css';
 
-const TodoItem = ({ deleteItem, toggleItem, isCompleted, text }) => (
+const TodoItem = ({
+  deleteItem, toggleItem, isCompleted, text,
+}) => (
   <li>
-    <input type="text"
+    <input
+      type="text"
       placeholder={text}
       defaultValue={text}
       className={(isCompleted)
-        ? 'todo-item-completed'
-        : 'todo-item-noncompleted'} >
+        ? styles.todoItemCompleted
+        : styles.todoItemNonCompleted}
+    >
     </input>
     <button type="submit" onClick={toggleItem}>Toggle</button>
     <button type="submit" onClick={deleteItem}>Delete</button>
@@ -21,7 +25,6 @@ TodoItem.propTypes = {
   toggleItem: PropTypes.func.isRequired,
   isCompleted: PropTypes.bool.isRequired,
   text: PropTypes.string.isRequired,
-  
 };
 
 export default TodoItem;
