+++ src/Components/TodoList/TodoList.module.css
@@ -1,0 +1,27 @@
+ul {
+	list-style-type: none;
+};
+
+.todo_list {
+	position: relative;
+	background-color: white;
+	margin: 30px;
+	width: 40%;
+	min-width: 400px;
+	max-width: 400px;
+	height: auto;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	border: 1px solid #efefef;
+	box-shadow:
+		0 1px 1px rgba(0,0,0,0.15),
+		0 10px 0 -5px #eee,
+		0 10px 1px -4px rgba(0,0,0,0.15),
+		0 20px 0 -10px #eee,
+		0 20px 1px -9px rgba(0,0,0,0.15);
+	padding: 30px;
+	overflow-y: auto;
+	scroll-snap-type: y mandatory;
+	scroll-behavior: smooth;
+};
